+++ CHANGELOG.md
@@ -113,6 +113,8 @@ Basic Authentication
 - username/password based login
 - uses bip39 for recovery phrases
 
+/additem
+- added ability to fuzzy search for matching name, or dbname
 ```
 
 ## 5.0.0